Initial schematic
This is the initial schematic for the door counter project.

Also have an incomplete BOM and some implementation ideas.

Decided on possible Arduino compatibility not because it is best for the
job, but because it is barely capable and others know how to work on it.
BOM is probably $2 higher and programming will probably be buggier an
take longer because of this. Still have time to determine if this should
be standard C or if the micro should be changed altogether.

Likewise, chose Eagle for CAD as many are familiar with it, and I don't
care for KiCad. (Don't care for Eagle much either, but...)